+++ source/Messaging.CimMessageAdapter/Messages/MessageParser.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Messaging.Domain.OutgoingMessages;
 
 namespace Messaging.CimMessageAdapter.Messages;
