+++ source/Messaging.CimMessageAdapter/Response/XmlResponseFactory.cs
@@ -17,6 +17,7 @@ using System.Linq;
 using System.Text;
 using System.Xml;
 using Messaging.CimMessageAdapter.Messages;
+using Messaging.Domain.OutgoingMessages;
 
 namespace Messaging.CimMessageAdapter.Response
 {
