+++ source/Messaging.Infrastructure/IncomingMessages/CimFormatParser.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Linq;
 using Messaging.CimMessageAdapter.Messages;
+using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.SeedWork;
 
 namespace Messaging.Infrastructure.IncomingMessages;
