+++ source/Messaging.Tests/Infrastructure/IncomingMessages/CimFormatParserTests.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Messaging.CimMessageAdapter.Messages;
+using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.SeedWork;
 using Messaging.Infrastructure.IncomingMessages;
 using Xunit;
